backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6
pyyaml==6.0.1
base64==1.0.0
urllib3==2.1.0
requests==2.31.0
jinja2==3.1.2